+++ packages/lib/modules/vebal/vote/Votes/MyVotes/myVotes.helpers.ts
@@ -1,6 +1,7 @@
 import { oneDayInMs, toJsTimestamp } from '@repo/lib/shared/utils/time'
 import { bn } from '@repo/lib/shared/utils/numbers'
 import { formatDistanceToNow } from 'date-fns'
+import { SortingBy } from './myVotes.types'
 
 export const WEIGHT_VOTE_DELAY = 10 * oneDayInMs
 
@@ -45,4 +46,23 @@ export function getUnallocatedWeight(weight: string | number) {
       .toNumber(),
     0
   )
+}
+
+export const orderByHash: Record<SortingBy, { label: string; title?: string }> = {
+  bribes: {
+    label: 'Bribes',
+    title:
+      'Voting incentives (referred to as ‘Bribes’ in DeFi) are provided by unaffiliated 3rd parties through the Hidden Hand platform to incentivize liquidity to certain pools.',
+  },
+  bribesPerVebal: {
+    label: 'Bribes/veBAL',
+    title:
+      'This shows the ratio of 3rd party voting incentives (known as Bribes) to veBAL. The higher this ratio, the more profitable it is to currently vote on this pool. Note this ratio could change up till the voting deadline.',
+  },
+  currentVotes: {
+    label: 'Current votes',
+  },
+  editVotes: {
+    label: 'Edit votes',
+  },
 }
